we refactored App.js to use GameLayout
