+++ yesnorpg/src/App.tsx
@@ -1,0 +1,7 @@
+import React, { ReactElement } from "react";
+import "./App.css";
+import { GameLayout } from "./pages/GameLayout";
+
+export function App(): ReactElement {
+  return <GameLayout />;
+}
